third_party/WebKit/LayoutTests/imported/web-platform-tests/cssom-view/elementFromPoint-expected.txt

 
 
    
Hello WPT!
 
Another teal
This is a testharness.js-based test.
PASS Negative co-ordinates 
PASS co-ordinates larger than the viewport 
PASS co-ordinates larger than the viewport from in iframe 
PASS Return first element that is the target for hit testing 
PASS First element to get mouse events with pointer-events css 
PASS SVG element at x,y 
FAIL transformed element at x,y assert_equals: Should have returned pink element that has a transform expected Element node <div id="pink" class="size pink" style="transform: transl... but got null
FAIL no hit target at x,y assert_equals: Should have returned the root expected Element node <body>
  <div id="purple" class="size purple">&nbsp;</div... but got null
PASS No viewport available 
Harness: the test ran to completion.
